requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.16
mysqlclient==2.1.1
uptime==3.0.1
requests==2.28.1
pyjwt==2.7.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.3.1
email_validator
flask_admin
Flask-Session
flask_bcrypt
psutil
flask-restful
python-dotenv
redis
flask_redis==0.4.0
